+++ src/main/java/io/zerogone/controller/api/UserSearchController.java
@@ -1,4 +1,4 @@
-package io.zerogone.controller;
+package io.zerogone.controller.api;
 
 import io.zerogone.model.UserVo;
 import io.zerogone.service.UserSearchService;
