+++ src/main/java/io/zerogone/service/UserSearchService.java
@@ -4,9 +4,6 @@ import io.zerogone.model.UserVo;
 import io.zerogone.repository.UserDao;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 @Service
 public class UserSearchService {
     private final UserDao userDao;
@@ -16,13 +13,6 @@ public class UserSearchService {
     }
 
     public UserVo getUserByEmail(String email) {
-        return new UserVo(userDao.findUserByEmail(email));
-    }
-
-    public List<UserVo> getUsersByBlogId(int blogId) {
-        return userDao.findAllByBlogId(blogId)
-                .stream()
-                .map(UserVo::new)
-                .collect(Collectors.toList());
+        return new UserVo(userDao.findByEmail(email));
     }
 }
